Improvements to Search function
Clear the search input after hitting an item by adding the search value to state and clearing it in the routeToItem function. Moved the routToItem function inside the component.

When doing a new search after routing to an item, the previous results was already loaded and shown briefly when doing a new search. Avoiding this by resetting the items array after route.

For a brief moment, the "Nothing found" was displayed when starting a search. Setting loading to true outside the debounce function solved this.